introduce new jwt.Parser
The current Parse function does not verify the alg, iss, and sub claims. By introducing a new Parser structure, it makes it easier to verify these claims.
