+++ jwt/key_finder.go
@@ -1,6 +1,7 @@
 package jwt
 
 import (
+	"context"
 	"errors"
 	"fmt"
 
@@ -9,22 +10,6 @@ import (
 	"github.com/shogo82148/goat/jws"
 	"github.com/shogo82148/goat/sig"
 )
-
-// KeyFinder is a wrapper for the FindKey method.
-type KeyFinder interface {
-	// FindKey finds the key used for signing.
-	// e.g, you can return a key corresponding to the KID.
-	FindKey(header *jws.Header) (key sig.SigningKey, err error)
-}
-
-// FindKeyFunc is an adapter to allow the use of ordinary functions as KeyFinder interfaces.
-// If f is a function with the appropriate signature, FindKeyFunc(f) is a KeyFinder that calls f.
-type FindKeyFunc func(header *jws.Header) (key sig.SigningKey, err error)
-
-// FindKey calls f(header).
-func (f FindKeyFunc) FindKey(header *jws.Header) (sig.SigningKey, error) {
-	return f(header)
-}
 
 // Token is a decoded JWT token.
 type Token struct {
@@ -56,7 +41,7 @@ type JWKKeyFiner struct {
 	Key *jwk.Key
 }
 
-func (f *JWKKeyFiner) FindKey(header *jws.Header) (sig.SigningKey, error) {
+func (f *JWKKeyFiner) FindKey(ctx context.Context, header *jws.Header) (sig.SigningKey, error) {
 	alg, err := guessAlg(f.Key, header)
 	if err != nil {
 		return nil, err
